feat(tests): Enhance tweet generation tests with output file logging
- Added TweetWriter utility to handle writing generated tweets to an output file.
- Implemented clearing of the output file at the start of tests to ensure fresh logs.
- Updated existing tests to log generated tweets for "Expired Proposal", "Expiring Soon Proposal", "New Proposal", and "New Vote Cast Tweet".
- Introduced a new support file for AgentTestSupport to facilitate test setup and proposal event generation.
