+++ .gitignore
@@ -31,6 +31,7 @@ dist-ssr
 
 build
 dist
+outputs
 *.tsbuildinfo
 .env
 *.db
